client/requirements.txt
# Requirements for HTTP MCP client with LangChain
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
langchain-core>=0.2.0
httpx>=0.25.0
opentelemetry-api>=1.20.0
pydantic>=2.0.0
